Update so that it really works on all operating systems
